Add rule for ball going out of field (due to chip kick) (both Div) (org-5)
From the Rule Change proposal:

Even with walls, the ball can still leave the field, if flying over the wall.
Teams should be encouraged to keep the ball inside the field, so there should be an adequate penalty for chipping the ball out.

If a teams chips the ball out, the foul counter of this team will be incremented.

If playing without walls, the game is restarted as if the ball left the field normally.
If playing with walls, the game is restarted with a free kick according to the current rules.

For simplicity, this change is not limited to games with walls.