Robot interchange during game play (Both Div) (org-9)
From the Rule proposal:

We aim for less stoppages and especially with walls, there will probably be much less chances to interchange robots. We already encourage teams to automatically interchange robots by replacing them at the midline, but the game still had to be stopped and even halted in 2019. This wasted overall game time.

We propose that robots can always be taken in and out from the midline. No need to stop the game. There should not be any action at the midline during the replacement. That could be enforced by requiring that the ball is not near the interchange position.

In 2019, a lot of time was lost in HALT, while teams interchanged their robots.
We want to encourage teams to reduce the number of manual interchanges to a minimum and make use of the automatic interchange to save overall game time.

We propose that manual robot interchange (if the robot is not taken in/out from the midline) always requires a timeout to be taken.
If a team is out of timeouts and still needs to take one robot out, the team is allowed to take the robot out without a timeout, but can not use any remaining timeout time. It can only remove the robot from the field as fast as possible.

Co-authored-by: Nicolai Ommer <nicolai.ommer@gmail.com>