Do not stop for cards (both Div) (org-10)
From the rule changes proposal:

With the introduction of the autoRef, a lot more fouls were detected then before. This also results in more yellow cards, which in turn stop the game.

We propose to not stop the game on yellow cards immediately, but allow teams within 10s to move a robot to the interchange position, where it can be taken out immediately.
This is especially useful for bot crashes, which increase the foul counter, which in turn may add yellow cards.
If a team did not get a robot out within time, the game is stopped and the ball is placed in front of the teams defense area and the other team gets a freekick.

A team can't score a goal while having more than the allowed number of robots on the field.

Co-authored-by: Nicolai Ommer <nicolai.ommer@gmail.com>