オンゲーム中のロボット交代 (#15)
Original : Pull Request 9

* Robot interchange during game play (Both Div)

From the Rule proposal:

We aim for less stoppages and especially with walls, there will probably be much less chances to interchange robots. We already encourage teams to automatically interchange robots by replacing them at the midline, but the game still had to be stopped and even halted in 2019. This wasted overall game time.

We propose that robots can always be taken in and out from the midline. No need to stop the game. There should not be any action at the midline during the replacement. That could be enforced by requiring that the ball is not near the interchange position.

In 2019, a lot of time was lost in HALT, while teams interchanged their robots.
We want to encourage teams to reduce the number of manual interchanges to a minimum and make use of the automatic interchange to save overall game time.

We propose that manual robot interchange (if the robot is not taken in/out from the midline) always requires a timeout to be taken.
If a team is out of timeouts and still needs to take one robot out, the team is allowed to take the robot out without a timeout, but can not use any remaining timeout time. It can only remove the robot from the field as fast as possible.

* 該当部分の日本語訳、リンクの修正

ルール変更案より:

我々はストップゲームの削減を目指しており、特に壁ありの試合では、
ロボットを交代させる機会がはるかに少なくなることが予想されます。
我々はすでに、各チームが自動的にロボットの交換を行うよう推奨していますが、
2019年の大会では依然として試合中にstop / haltが行われました。
これはゲーム全体の時間を無駄にしました。

そこで我々は、ミッドラインからであれば常時ロボットを交代できることを提案します。
ゲームを止める必要はありません。
ただし交代中に、ミッドライン付近でいかなるアクションも行われるべきではありません。
これは、ボールが交換位置の付近に無いことを条件とすることで実施できます。

2019年、ロボット交換のため多くの時間がhaltに費やされました。
各チームが手動でのロボット交代の数を最小限に減らし、自動交代を利用して
ゲーム全体の時間を節約することを推奨します。

手動によるロボット交代(ロボットがミッドラインから出し入れされない場合)は、
常にタイムアウトを取得するようにすることを提案します。
仮にチームがタイムアウトを使い切っていて、かつロボットを出さなければならない場合、
タイムアウトを取得せずロボットを出すことができますが、残りのタイムアウト時間は使えません。
可能な限り早く、ロボットをフィールドから取り除くことだけができます。

Translated by: Kazunori Kimura

Co-authored-by: Leonardo Costa <37155964+leonardoSCosta@users.noreply.github.com>
Co-authored-by: Nicolai Ommer <nicolai.ommer@gmail.com>