Rules2020/14 ペナルティーキックの削除 (#19)
* Remove Penalty Kicks (both Div) (org-14)

From the rule proposal:

In 2019, there are two ways to receive a penalty kick:

Every third card led to a penalty kick
If a defender other than the keeper is fully inside the defense area and touches the ball
Penalty kicks have several disadvantages:

They are quite unbalanced (and thus not very interesting)
They lack innovation
They need extra implementation
They require additional time during a match
They interrupt the game
We propose to get rid of penalty kicks completely:

The Multiple Defenders foul will result in a red card. The game will be stopped immediately.
The third active yellow card will be turned into a red card instead. A yellow card is active as long as the counter has not run out, so in order to get a red card, the team must collect three yellow cards within 2 minutes.
During the last team meeting, it was suggested to replace the current penalty kick by the one-on-one procedure that is currently only applied in the shoot-out phase.
This is not as hard of a punishment as the current penalty kick, but it was found a better punishment than the red card.

This PR replaces the penalty kick procedure by the one-on-one procedure. There are no changes in what causes a penalty kick in this PR.

* 該当部分の日本語訳、リンク修正

ルール変更案より:

2019年のルールでは、ペナルティーキックに至る２種類の場合がありました。

(イエロー)カードが3枚たまるとペナルティーキックです。
ディフェンダーが自陣ディフェンスエリアに完全に入った状態でボールに接触した場合、これもペナルティーキックとなります。
ペナルティーキックには、複数の問題点があります:

アンバランスです(そして、たいして面白くありません)
革新性に欠けます
個別に実装しなくてはなりません
試合中にさらなる時間を設けなければなりません
試合に割り込み、中断させます

我々はペナルティーキックの撤廃を提案します:

マルチプルディフェンスはレッドカードとします。試合は即座に停止されます。
三枚目のイエローカードは、代わりにレッドカードとします。イエローカードのカウンターは健在ですので、レッドカードを得るには、チームは2分以内に3枚のイエローカードを得なければなりません。
前回のチームミーティングにおいて、現在のペナルティーキックを、ペナルティーシュートアウトでのみ用いられているようなone-on-oneに置き換えることが提案されました。
これは、現在のペナルティキックほど厳しい罰ではありませんが、レッドカードよりも優れた罰であるように思えます。

このPRはペナルティーキックをone-on-oneに置き換えます。
ペナルティーキックとなる理由に変更はありません。

Translated by: Kazunori Kimura

Co-authored-by: Nicolai Ommer <nicolai.ommer@gmail.com>