Minor fixes and updates (origin-35)
* Clarification of robot substitution when putting robots in

It is possible to temporally exceed the number of robots now,
e.g. through yellow cards.
The 'any time' should be related to putting robots in.

* Reset timer on event only for Defender Too Close To Ball

This makes it easier to handle in the autoRefs and GC and should
not make a significant difference in practice.

* Fix a spelling mistake

* Add the possible goal event to the appendix

* Add missing link to free kick section

* Update game event tables in appendix

* Keeper held ball event is triggered by GC

Minor fixes and updates (origin-35)

* Clarification of robot substitution when putting robots in

It is possible to temporally exceed the number of robots now,
e.g. through yellow cards.
The 'any time' should be related to putting robots in.

* Reset timer on event only for Defender Too Close To Ball

This makes it easier to handle in the autoRefs and GC and should
not make a significant difference in practice.

* Fix a spelling mistake

* Add the possible goal event to the appendix

* Add missing link to free kick section

* Update game event tables in appendix

* Keeper held ball event is triggered by GC